refactor(#283): split animation, timer hook in ChatInput component
